+++ tests/DbEx.Test.Console/Migrations/004-create-test-contact-table.sql
@@ -4,6 +4,7 @@
       [Phone] VARCHAR (15) NULL,
       [DateOfBirth] DATE NULL,
       [ContactTypeId] INT NOT NULL DEFAULT 1,
-      [GenderId] INT NULL
+      [GenderId] INT NULL,
+      [TenantId] NVARCHAR(50),
       CONSTRAINT [FK_Test_Contact_ContactType] FOREIGN KEY ([ContactTypeId]) REFERENCES [Test].[ContactType] ([ContactTypeId])
     )